feat: 自研高性能 JSON 解析器 + 修复工具调用参数损坏 + 修复中文输入乱码
新增 container::Json 解析器:
- 递归下降解析器，所有权字符串模式（非零拷贝，避免悬空）
- 两遍序列化器（预计算大小 + 单次写入）
- SIMD 加速抽象层（SSE4.2/AVX2/NEON/Scalar 运行时调度）
- ProxyRef 链式访问 j["a"]["b"]["c"] = val 零拷贝写入 DOM
- API 兼容 nlohmann/json，业务代码零修改
- JsonObject 开放寻址法 + JsonArray 连续内存
- JsonValue 拷贝构造自动升级零拷贝为所有权（防御性）

修复工具调用参数损坏:
- 根因：Json::parse() 零拷贝字符串指向临时输入缓冲区，缓冲区销毁后悬空
- 修复：Parser 直接创建所有权字符串，消除零拷贝
- 修复：JsonValue 拷贝构造自动升级零拷贝为所有权
- 修复：ProxyRef::begin()/end() 直接从 DOM 节点构造迭代器，避免临时 Json 悬空
- 优化：ensure_all_owned() 改为迭代实现（栈深度 256），防止深层 JSON 栈溢出

修复中文输入乱码:
- 根因：macOS IME 偶发 ESC + UTF-8 首字节，parse_escape() 消费掉 UTF-8 首字节
- 修复：TerminalIO 添加 pushback 缓冲区，parse_escape() 检测 UTF-8 字节后放回
- 修复：read_utf8_char() 续字节失败时返回空字符串，不插入残缺 UTF-8

测试: 283 tests passed

+++ include/ben_gear/tools/workflow_tools.hpp
@@ -103,7 +103,7 @@ inline void register_workflow_tools_with_resources(
                         task.type != "subflow") {
                         Json error_result;
                         error_result["success"] = false;
-                        error_result["error"] = "Invalid task type: " + task.type + 
+                        error_result["error"] = "Invalid task type: " + task.type +
                             ". Supported types: llm, tool, function, condition, subflow";
                         return container::String(error_result.dump().c_str());
                     }
@@ -283,11 +283,11 @@ inline void register_workflow_tools_with_resources(
                 
                 result["progress"]["completed"] = completed;
                 result["progress"]["failed"] = failed;
-                result["progress"]["total"] = state->task_results.size();
+                result["progress"]["total"] = static_cast<uint64_t>(state->task_results.size());
                 
                 if (state->task_results.size() > 0) {
                     result["progress"]["percentage"] = 
-                        (completed * 100) / state->task_results.size();
+                        (static_cast<uint64_t>(completed) * 100) / static_cast<uint64_t>(state->task_results.size());
                 }
                 
                 // 任务状态列表
